# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API provided by DTI project",
        "title": "Domain Threat Intelligence API",
        "contact": {
            "name": "Yaroslav Lysak",
            "url": "https://t.me/Qvineox"
        },
        "version": "0.0.4"
    },
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Accepts login and password, return pair of auth tokens",
                "tags": [
                    "Auth"
                ],
                "summary": "Authorizes user by login and password",
                "parameters": [
                    {
                        "description": "user credentials",
                        "name": "username",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routing.loginParams"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/routing.tokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Removes auth info from cookie and database by provided refresh token",
                "tags": [
                    "Auth"
                ],
                "summary": "Removes user auth tokens",
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/routing.tokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    }
                }
            }
        },
        "/auth/password-strength": {
            "post": {
                "description": "Returns password strength",
                "tags": [
                    "Auth"
                ],
                "summary": "Get strength of a password",
                "parameters": [
                    {
                        "description": "password",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routing.passwordStrengthParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routing.passwordStrengthResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Gets refresh token from Cookie and updates user auth tokens",
                "tags": [
                    "Auth"
                ],
                "summary": "Updates user auth tokens",
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/routing.tokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    }
                }
            }
        },
        "/auth/registration": {
            "post": {
                "description": "Accepts user account data and register new platform user",
                "tags": [
                    "Auth"
                ],
                "summary": "Creates new user with defined data",
                "parameters": [
                    {
                        "description": "user credentials",
                        "name": "username",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routing.loginParams"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "error.APIError": {
            "type": "object",
            "properties": {
                "ErrorCode": {
                    "$ref": "#/definitions/error.CodeError"
                },
                "ErrorMessage": {
                    "type": "string"
                },
                "ErrorModule": {
                    "type": "string"
                },
                "StatusCode": {
                    "type": "integer"
                }
            }
        },
        "error.CodeError": {
            "type": "integer",
            "enum": [
                1,
                2,
                3,
                4,
                5,
                6,
                7,
                8,
                9,
                10,
                11,
                12,
                13,
                14,
                15
            ],
            "x-enum-varnames": [
                "EncodingErrorCode",
                "DecodingErrorCode",
                "IncorrectParamsErrorCode",
                "InsufficientParamsErrorCode",
                "DatabaseErrorCode",
                "DatabaseEntityNotFoundCode",
                "FileDecodingErrorCode",
                "FileReadingErrorCode",
                "FileExtensionNotSupportedErrorCode",
                "NotImplementedErrorCode",
                "FileProcessingErrorCode",
                "InternalUnidentifiedErrorCode",
                "AuthFailedErrorCode",
                "AuthPermissionInsufficientErrorCode",
                "JobQueueErrorCode"
            ]
        },
...
        "routing.loginParams": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "routing.passwordStrengthParams": {
            "type": "object",
            "required": [
                "Password"
            ],
            "properties": {
                "Password": {
                    "type": "string"
                }
            }
        },
        "routing.passwordStrengthResponse": {
            "type": "object",
            "properties": {
                "CrackTime": {
                    "type": "number"
                },
                "Entropy": {
                    "type": "number"
                },
                "Level": {
                    "type": "integer"
                }
            }
        },
...
        "routing.tokenResponse": {
            "type": "object",
            "properties": {
                "AccessToken": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "x-api-Key",
            "in": "header"
        }
    }
}